BunchOfStuff
i added object to spawn in the two basic units for each side on
specified tiles
i created a system to pick one unit by clicking on the tile it is
located on with the left mouse button and using the wasd keys you can
move it around.
if you wish to move a different unity all you need to do is click on the
tile it is located on, this works for both white units and black units.
i think that's all i did so far but i'm not sure.