UpdatedMovementControls
I've changed the movement controls for the pieces on the board. Simply
click on whichever piece you wish to move and the tile you hover over
next will show if you may place that piece there with either green or
red. Next simply flick on the tile you wish to move it to and it will be
moved there.