ScenesConnectedV1
the scenes TestingHexGrid, and BattleTest are now connected, when you go
into a fight you get put into the fight and after someone loses their
piece is destroyed and the winners takes the tile!
This is not dependent on which piece you are controlling but the piece
that initiates combat is player01 and the defending player is player02.