PieceCounterV1.0
I've implemented a system to display how many pieces of each type each
player has in play. The sprites i'm using at the moment are the health
bar hex's but those sprites can be later created. Green is for Grunts,
Red is for Tanks and Blue is for Scouts. Right now there is a limit to
only 4 units of each type but this may be changed. Once the Piece is
destroyed in combat, that piece's sprite is also deactivated.