Merge remote-tracking branch 'origin/master'
Conflicts:
	ArchonClone/Assets/TestScenes/TestingHexTiles.unity

Merge remote-tracking branch
'origin/master'

Conflicts:
	ArchonClone/Assets/TestScenes/TestingHexTiles.unity

New
Movement System!!!!!

Now when you select a piece it checks if it
takes
too many nodes or
tiles to get to that piece. If it takes too
many; even
if it is within
your pieces movement radius(which I am trying
to update
to show it is a
bad move) it will not allow the player to move
to that
tile! also fixed
a bug that I might have made while implementing
new
movement system!